♻️ Flytt endepunkt for å sjekke om brker har remmevedtak til vedtakControlleren, da det hører mer hjemme her enn i vilkårControlleren
