+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/dagligReise/DagligReiseVedtakController.kt
@@ -1,6 +1,7 @@
 package no.nav.tilleggsstonader.sak.vedtak.dagligReise
 
 import no.nav.security.token.support.core.api.ProtectedWithClaims
+import no.nav.tilleggsstonader.kontrakter.felles.Stønadstype
 import no.nav.tilleggsstonader.sak.behandling.BehandlingService
 import no.nav.tilleggsstonader.sak.behandling.domain.Behandling
 import no.nav.tilleggsstonader.sak.behandlingsflyt.StegFerdigstiltResponse
@@ -16,6 +17,7 @@ import no.nav.tilleggsstonader.sak.vedtak.VedtakService
 import no.nav.tilleggsstonader.sak.vedtak.dagligReise.beregning.DagligReiseBeregningService
 import no.nav.tilleggsstonader.sak.vedtak.dagligReise.dto.AvslagDagligReiseDto
 import no.nav.tilleggsstonader.sak.vedtak.dagligReise.dto.BeregningDagligReiseDto
+import no.nav.tilleggsstonader.sak.vedtak.dagligReise.dto.HarRammevedtakDto
 import no.nav.tilleggsstonader.sak.vedtak.dagligReise.dto.InnvilgelseDagligReiseTsoRequest
 import no.nav.tilleggsstonader.sak.vedtak.dagligReise.dto.InnvilgelseDagligReiseTsrRequest
 import no.nav.tilleggsstonader.sak.vedtak.dagligReise.dto.OpphørDagligReiseRequest
@@ -47,7 +49,30 @@ class DagligReiseVedtakController(
     private val validerGyldigÅrsakAvslag: ValiderGyldigÅrsakAvslag,
     private val beregningsplanUtleder: BeregningsplanUtleder,
     private val dagligReiseVilkårService: DagligReiseVilkårService,
+    private val dagligReiseVedtakService: DagligReiseVedtakService,
 ) {
+    @GetMapping("{behandlingId}/har-rammevedtak")
+    fun harRammevedtak(
+        @PathVariable behandlingId: BehandlingId,
+    ): HarRammevedtakDto {
+        tilgangService.settBehandlingsdetaljerForRequest(behandlingId)
+        tilgangService.validerLesetilgangTilBehandling(behandlingId)
+
+        val saksbehandling = behandlingService.hentSaksbehandling(behandlingId)
+
+        if (saksbehandling.stønadstype !in setOf(Stønadstype.DAGLIG_REISE_TSO, Stønadstype.DAGLIG_REISE_TSR)) {
+            return HarRammevedtakDto(false)
+        }
+
+        val harRammevedtak =
+            dagligReiseVedtakService.harRammevedtakPåDenneEllerForrgieBehandling(
+                behandlingId = saksbehandling.id,
+                forrigeIverksatteBehandlingId = saksbehandling.forrigeIverksatteBehandlingId,
+            )
+
+        return HarRammevedtakDto(harRammevedtak)
+    }
+
     @PostMapping("{behandlingId}/tso/innvilgelse")
     fun innvilge(
         @PathVariable behandlingId: BehandlingId,
