+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/dagligReise/dto/HarRammevedtakDto.kt
@@ -1,0 +1,5 @@
+package no.nav.tilleggsstonader.sak.vedtak.dagligReise.dto
+
+data class HarRammevedtakDto(
+    val harRammevedtak: Boolean,
+)
