--- src/main/kotlin/no/nav/tilleggsstonader/sak/vilkår/stønadsvilkår/dagligReise/dto/HarRammevedtakDto.kt
@@ -1,5 +1,0 @@
-package no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.dagligReise.dto
-
-data class HarRammevedtakDto(
-    val harRammevedtak: Boolean,
-)
